fix clientside error with types
