+++ packages/db/src/index.js
@@ -1,18 +1,4 @@
 "use strict";
-var __createBinding = (this && this.__createBinding) || (Object.create ? (function(o, m, k, k2) {
-    if (k2 === undefined) k2 = k;
-    var desc = Object.getOwnPropertyDescriptor(m, k);
-    if (!desc || ("get" in desc ? !m.__esModule : desc.writable || desc.configurable)) {
-      desc = { enumerable: true, get: function() { return m[k]; } };
-    }
-    Object.defineProperty(o, k2, desc);
-}) : (function(o, m, k, k2) {
-    if (k2 === undefined) k2 = k;
-    o[k2] = m[k];
-}));
-var __exportStar = (this && this.__exportStar) || function(m, exports) {
-    for (var p in m) if (p !== "default" && !Object.prototype.hasOwnProperty.call(exports, p)) __createBinding(exports, m, p);
-};
 Object.defineProperty(exports, "__esModule", { value: true });
 exports.db = void 0;
 const serverless_1 = require("@neondatabase/serverless");
@@ -53,7 +39,6 @@ const createPrismaClient = () => {
     throw new Error("Database adapter not configured for this environment");
 };
 const globalForPrisma = globalThis;
-__exportStar(require("@prisma/client"), exports);
 exports.db = globalForPrisma.prisma ?? createPrismaClient();
 if (process.env.NODE_ENV !== "production")
     globalForPrisma.prisma = exports.db;
