+++ packages/db/src/types.ts
@@ -1,0 +1,1 @@
+export * from "@prisma/client";
